+++ tests/https/test_unit_http.py
@@ -12,7 +12,7 @@ from asynctest import patch
 
 from aiohttp import BasicAuth
 
-from conjur_api.models import SslVerificationMode, SslVerificationMetadata
+from conjur_api.models import SslVerificationMode, SslVerificationMetadata, ProxyParams
 from conjur_api.errors.errors import HttpSslError
 from conjur_api.wrappers.http_wrapper import HttpVerb, invoke_endpoint
 from tests.https.common import MockResponse
@@ -58,7 +58,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=None, headers={}, data='', ssl=ssl_context,
-                                                 params=None)
+                                                 params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_can_handle_unset_params(self, mock_request):
@@ -69,7 +69,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=None, headers={}, data='', ssl=ssl_context,
-                                                 params=None)
+                                                 params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_uses_http_verb_for_method_name(self, mock_request):
@@ -79,15 +79,15 @@ class HttpInvokeEndpointTest(AsyncTestCase):
             await invoke_endpoint(HttpVerb.GET, self.MockEndpoint.NO_PARAMS, {})
             mock_create_ssl_context.assert_called_with()
             mock_request.assert_called_with('GET', 'no/params', auth=None, headers={}, data='', ssl=ssl_context,
-                                            params=None)
+                                            params=None, proxy=None)
 
             await invoke_endpoint(HttpVerb.POST, self.MockEndpoint.NO_PARAMS, {})
             mock_request.assert_called_with('POST', 'no/params', auth=None, headers={}, data='', ssl=ssl_context,
-                                            params=None)
+                                            params=None, proxy=None)
 
             await invoke_endpoint(HttpVerb.DELETE, self.MockEndpoint.NO_PARAMS, {})
             mock_request.assert_called_with('DELETE', 'no/params', auth=None, headers={}, data='', ssl=ssl_context,
-                                            params=None)
+                                            params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_generates_url_from_endpoint_object(self, mock_request):
@@ -98,7 +98,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'http://host/no/params', auth=None, headers={}, data='',
-                                                 ssl=ssl_context, params=None)
+                                                 ssl=ssl_context, params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_attaches_api_token_header_if_present_in_params(self, mock_request):
@@ -109,7 +109,8 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=None, data='', ssl=ssl_context,
-                                                 headers={'Authorization': 'Token token="dG9rZW4="'}, params=None)
+                                                 headers={'Authorization': 'Token token="dG9rZW4="'}, params=None,
+                                                 proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_verifies_ssl_by_default(self, mock_request):
@@ -120,7 +121,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=None, data='', ssl=ssl_context, headers={},
-                                                 params=None)
+                                                 params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_ssl_verify_param_defaults_to_true_to_http_client(self, mock_request):
@@ -132,7 +133,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=None, data='', ssl=ssl_context, headers={},
-                                                 params=None)
+                                                 params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_raises_hostname_mismatch_error(self, mock_request):
@@ -149,7 +150,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
             ssl_context_calls = [call(cafile='foo')]
             mock_create_ssl_context.assert_has_calls(ssl_context_calls)
             mock_request.assert_called_with('GET', 'no/params', auth=None, data='', ssl=ssl_context, headers={},
-                                            params=None)
+                                            params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_passes_auth_param_to_http_client_if_provided(self, mock_request):
@@ -160,7 +161,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=BasicAuth('foo', 'bar'), data='',
-                                                 ssl=ssl_context, headers={}, params=None)
+                                                 ssl=ssl_context, headers={}, params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_passes_extra_args_to_http_client(self, mock_request):
@@ -171,7 +172,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=None, data='ab', ssl=ssl_context,
-                                                 headers={}, params=None)
+                                                 headers={}, params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_passes_query_param(self, mock_request):
@@ -186,7 +187,19 @@ class HttpInvokeEndpointTest(AsyncTestCase):
 
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', 'no/params', auth=None, data='ab', ssl=ssl_context,
-                                                 headers={}, params=query)
+                                                 headers={}, params=query, proxy=None)
+
+    @patch('aiohttp.ClientSession.request')
+    async def test_invoke_endpoint_passes_proxy_param(self, mock_request):
+        ssl_context = ssl.create_default_context()
+        with patch.object(ssl, 'create_default_context', return_value=ssl_context) as mock_create_ssl_context:
+            mock_request.return_value = MockResponse('', 200)
+            await invoke_endpoint(HttpVerb.GET, self.MockEndpoint.NO_PARAMS, None, 'ab',
+                                  proxy_params=ProxyParams('proxy.com'))
+
+            mock_create_ssl_context.assert_called_once_with()
+            mock_request.assert_called_once_with('GET', 'no/params', auth=None, data='ab', ssl=ssl_context,
+                                                 headers={}, params=None, proxy='proxy.com')
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_quotes_all_params_except_url(self, mock_request):
@@ -199,7 +212,7 @@ class HttpInvokeEndpointTest(AsyncTestCase):
             quoted_endpoint = '/'.join([self.UNESCAPED_PARAMS['url']] + self.ESCAPED_PARAMS)
             mock_create_ssl_context.assert_called_once_with()
             mock_request.assert_called_once_with('GET', quoted_endpoint, data='$#\\% ^%', auth=None,
-                                                 ssl=ssl_context, headers={}, params=None)
+                                                 ssl=ssl_context, headers={}, params=None, proxy=None)
 
     @patch('aiohttp.ClientSession.request')
     async def test_invoke_endpoint_raises_error_if_bad_status_code_is_returned(self, mock_request):